message as first argument, whole object as second
It seems like a good idea to read out other things then just the message. Backwards compatibility is ensured by still sending the message as a first parameter.

In these examples: https://developers.google.com/cloud-messaging/downstream#receiving-messages-on-an-android-client-app 
it would seem that that is the intention of the message.

I'm not quite sure if this will work though, because I couldn't find a 'how-to-contribute' kind of page. So I'm not sure how to build the `.jar` that's a part of the npm package.

Let me know what you think of this change.